# Conflicts:
#	CHANGELOG.md
#	composer.json
#	src/Illuminate/Foundation/Application.php
#	src/Illuminate/Log/composer.json

+++ src/Illuminate/Database/Schema/Grammars/SQLiteGrammar.php
@@ -5,8 +5,6 @@ namespace Illuminate\Database\Schema\Grammars;
 use Illuminate\Database\Connection;
 use Illuminate\Database\Query\Expression;
 use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Database\Schema\ColumnDefinition;
-use Illuminate\Database\Schema\ForeignKeyDefinition;
 use Illuminate\Database\Schema\IndexDefinition;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Fluent;
@@ -29,6 +27,23 @@ class SQLiteGrammar extends Grammar
     protected $serials = ['bigInteger', 'integer', 'mediumInteger', 'smallInteger', 'tinyInteger'];
 
     /**
+     * Get the commands to be compiled on the alter command.
+     *
+     * @param  \Illuminate\Database\Connection  $connection
+     * @return array
+     */
+    public function getAlterCommands(Connection $connection)
+    {
+        $alterCommands = ['change', 'primary', 'dropPrimary', 'foreign', 'dropForeign'];
+
+        if (version_compare($connection->getServerVersion(), '3.35', '<')) {
+            $alterCommands[] = 'dropColumn';
+        }
+
+        return $alterCommands;
+    }
+
+    /**
      * Compile the query to determine the SQL text that describes the given object.
      *
      * @param  string  $name
@@ -38,8 +53,8 @@ class SQLiteGrammar extends Grammar
     public function compileSqlCreateStatement($name, $type = 'table')
     {
         return sprintf('select "sql" from sqlite_master where type = %s and name = %s',
-            $this->wrap($type),
-            $this->wrap(str_replace('.', '__', $name))
+            $this->quoteString($type),
+            $this->quoteString(str_replace('.', '__', $name))
         );
     }
 
@@ -91,7 +106,7 @@ class SQLiteGrammar extends Grammar
         return sprintf(
             'select name, type, not "notnull" as "nullable", dflt_value as "default", pk as "primary", hidden as "extra" '
             .'from pragma_table_xinfo(%s) order by cid asc',
-            $this->wrap(str_replace('.', '__', $table))
+            $this->quoteString(str_replace('.', '__', $table))
         );
     }
 
@@ -104,12 +119,12 @@ class SQLiteGrammar extends Grammar
     public function compileIndexes($table)
     {
         return sprintf(
-            'select "primary" as name, group_concat(col) as columns, 1 as "unique", 1 as "primary" '
+            'select \'primary\' as name, group_concat(col) as columns, 1 as "unique", 1 as "primary" '
             .'from (select name as col from pragma_table_info(%s) where pk > 0 order by pk, cid) group by name '
-            .'union select name, group_concat(col) as columns, "unique", origin = "pk" as "primary" '
+            .'union select name, group_concat(col) as columns, "unique", origin = \'pk\' as "primary" '
             .'from (select il.*, ii.name as col from pragma_index_list(%s) il, pragma_index_info(il.name) ii order by il.seq, ii.seqno) '
             .'group by name, "unique", "primary"',
-            $table = $this->wrap(str_replace('.', '__', $table)),
+            $table = $this->quoteString(str_replace('.', '__', $table)),
             $table
         );
     }
@@ -127,7 +142,7 @@ class SQLiteGrammar extends Grammar
             .'group_concat("to") as foreign_columns, on_update, on_delete '
             .'from (select * from pragma_foreign_key_list(%s) order by id desc, seq) '
             .'group by id, "table", on_update, on_delete',
-            $this->wrap(str_replace('.', '__', $table))
+            $this->quoteString(str_replace('.', '__', $table))
         );
     }
 
@@ -214,15 +229,72 @@ class SQLiteGrammar extends Grammar
      *
      * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
      * @param  \Illuminate\Support\Fluent  $command
-     * @return array
+     * @return string
      */
     public function compileAdd(Blueprint $blueprint, Fluent $command)
     {
-        $columns = $this->prefixArray('add column', $this->getColumns($blueprint));
+        return sprintf('alter table %s add column %s',
+            $this->wrapTable($blueprint),
+            $this->getColumn($blueprint, $command->column)
+        );
+    }
 
-        return collect($columns)->map(function ($column) use ($blueprint) {
-            return 'alter table '.$this->wrapTable($blueprint).' '.$column;
-        })->all();
+    /**
+     * Compile alter table command into a series of SQL statements.
+     *
+     * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
+     * @param  \Illuminate\Support\Fluent  $command
+     * @param  \Illuminate\Database\Connection  $connection
+     * @return array|string
+     *
+     * @throws \RuntimeException
+     */
+    public function compileAlter(Blueprint $blueprint, Fluent $command, Connection $connection)
+    {
+        $columnNames = [];
+        $autoIncrementColumn = null;
+
+        $columns = collect($blueprint->getState()->getColumns())
+            ->map(function ($column) use ($blueprint, &$columnNames, &$autoIncrementColumn) {
+                $name = $this->wrap($column);
+
+                $autoIncrementColumn = $column->autoIncrement ? $column->name : $autoIncrementColumn;
+
+                if (is_null($column->virtualAs) && is_null($column->virtualAsJson) &&
+                    is_null($column->storedAs) && is_null($column->storedAsJson)) {
+                    $columnNames[] = $name;
+                }
+
+                return $this->addModifiers(
+                    $this->wrap($column).' '.($column->full_type_definition ?? $this->getType($column)),
+                    $blueprint,
+                    $column
+                );
+            })->all();
+
+        $indexes = collect($blueprint->getState()->getIndexes())
+            ->reject(fn ($index) => str_starts_with('sqlite_', $index->index))
+            ->map(fn ($index) => $this->{'compile'.ucfirst($index->name)}($blueprint, $index))
+            ->all();
+
+        $tempTable = $this->wrap('__temp__'.$blueprint->getPrefix().$blueprint->getTable());
+        $table = $this->wrapTable($blueprint);
+        $columnNames = implode(', ', $columnNames);
+
+        $foreignKeyConstraintsEnabled = $connection->scalar('pragma foreign_keys');
+
+        return array_filter(array_merge([
+            $foreignKeyConstraintsEnabled ? $this->compileDisableForeignKeyConstraints() : null,
+            sprintf('create table %s (%s%s%s)',
+                $tempTable,
+                implode(', ', $columns),
+                $this->addForeignKeys($blueprint->getState()->getForeignKeys()),
+                $autoIncrementColumn ? '' : $this->addPrimaryKeys($blueprint->getState()->getPrimaryKey())
+            ),
+            sprintf('insert into %s (%s) select %s from %s', $tempTable, $columnNames, $columnNames, $table),
+            sprintf('drop table %s', $table),
+            sprintf('alter table %s rename to %s', $tempTable, $table),
+        ], $indexes, [$foreignKeyConstraintsEnabled ? $this->compileEnableForeignKeyConstraints() : null]));
     }
 
     /**
@@ -237,94 +309,19 @@ class SQLiteGrammar extends Grammar
      */
     public function compileChange(Blueprint $blueprint, Fluent $command, Connection $connection)
     {
-        $schema = $connection->getSchemaBuilder();
-        $table = $blueprint->getTable();
+        // Handled on table alteration...
+    }
 
-        $changedColumns = collect($blueprint->getChangedColumns());
-        $columnNames = [];
-        $autoIncrementColumn = null;
-
-        $columns = collect($schema->getColumns($table))
-            ->map(function ($column) use ($blueprint, $changedColumns, &$columnNames, &$autoIncrementColumn) {
-                $column = $changedColumns->first(fn ($col) => $col->name === $column['name'], $column);
-
-                if ($column instanceof Fluent) {
-                    $name = $this->wrap($column);
-                    $autoIncrementColumn = $column->autoIncrement ? $column->name : $autoIncrementColumn;
-
-                    if (is_null($column->virtualAs) && is_null($column->virtualAsJson) &&
-                        is_null($column->storedAs) && is_null($column->storedAsJson)) {
-                        $columnNames[] = $name;
-                    }
-
-                    return $this->addModifiers($name.' '.$this->getType($column), $blueprint, $column);
-                } else {
-                    $name = $this->wrap($column['name']);
-                    $autoIncrementColumn = $column['auto_increment'] ? $column['name'] : $autoIncrementColumn;
-                    $isGenerated = ! is_null($column['generation']);
-
-                    if (! $isGenerated) {
-                        $columnNames[] = $name;
-                    }
-
-                    return $this->addModifiers($name.' '.$column['type'], $blueprint,
-                        new ColumnDefinition([
-                            'change' => true,
-                            'type' => $column['type_name'],
-                            'nullable' => $column['nullable'],
-                            'default' => $column['default'] ? new Expression($column['default']) : null,
-                            'autoIncrement' => $column['auto_increment'],
-                            'collation' => $column['collation'],
-                            'comment' => $column['comment'],
-                            'virtualAs' => $isGenerated && $column['generation']['type'] === 'virtual'
-                                ? $column['generation']['expression'] : null,
-                            'storedAs' => $isGenerated && $column['generation']['type'] === 'stored'
-                                ? $column['generation']['expression'] : null,
-                        ])
-                    );
-                }
-            })->all();
-
-        $foreignKeys = collect($schema->getForeignKeys($table))->map(fn ($foreignKey) => new ForeignKeyDefinition([
-            'columns' => $foreignKey['columns'],
-            'on' => $foreignKey['foreign_table'],
-            'references' => $foreignKey['foreign_columns'],
-            'onUpdate' => $foreignKey['on_update'],
-            'onDelete' => $foreignKey['on_delete'],
-        ]))->all();
-
-        [$primary, $indexes] = collect($schema->getIndexes($table))->map(fn ($index) => new IndexDefinition([
-            'name' => match (true) {
-                $index['primary'] => 'primary',
-                $index['unique'] => 'unique',
-                default => 'index',
-            },
-            'index' => $index['name'],
-            'columns' => $index['columns'],
-        ]))->partition(fn ($index) => $index->name === 'primary');
-
-        $indexes = collect($indexes)->reject(fn ($index) => str_starts_with('sqlite_', $index->index))->map(
-            fn ($index) => $this->{'compile'.ucfirst($index->name)}($blueprint, $index)
-        )->all();
-
-        $tempTable = $this->wrap('__temp__'.$blueprint->getPrefix().$table);
-        $table = $this->wrapTable($blueprint);
-        $columnNames = implode(', ', $columnNames);
-
-        $foreignKeyConstraintsEnabled = $connection->scalar('pragma foreign_keys');
-
-        return array_filter(array_merge([
-            $foreignKeyConstraintsEnabled ? $this->compileDisableForeignKeyConstraints() : null,
-            sprintf('create table %s (%s%s%s)',
-                $tempTable,
-                implode(', ', $columns),
-                $this->addForeignKeys($foreignKeys),
-                $autoIncrementColumn ? '' : $this->addPrimaryKeys($primary->first())
-            ),
-            sprintf('insert into %s (%s) select %s from %s', $tempTable, $columnNames, $columnNames, $table),
-            sprintf('drop table %s', $table),
-            sprintf('alter table %s rename to %s', $tempTable, $table),
-        ], $indexes, [$foreignKeyConstraintsEnabled ? $this->compileEnableForeignKeyConstraints() : null]));
+    /**
+     * Compile a primary key command.
+     *
+     * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
+     * @param  \Illuminate\Support\Fluent  $command
+     * @return string
+     */
+    public function compilePrimary(Blueprint $blueprint, Fluent $command)
+    {
+        // Handled on table creation or alteration...
     }
 
     /**
@@ -382,7 +379,7 @@ class SQLiteGrammar extends Grammar
      */
     public function compileForeign(Blueprint $blueprint, Fluent $command)
     {
-        // Handled on table creation...
+        // Handled on table creation or alteration...
     }
 
     /**
@@ -445,15 +442,33 @@ class SQLiteGrammar extends Grammar
      * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
      * @param  \Illuminate\Support\Fluent  $command
      * @param  \Illuminate\Database\Connection  $connection
-     * @return array
+     * @return array|null
      */
     public function compileDropColumn(Blueprint $blueprint, Fluent $command, Connection $connection)
     {
+        if (version_compare($connection->getServerVersion(), '3.35', '<')) {
+            // Handled on table alteration...
+
+            return null;
+        }
+
         $table = $this->wrapTable($blueprint);
 
         $columns = $this->prefixArray('drop column', $this->wrapArray($command->columns));
 
         return collect($columns)->map(fn ($column) => 'alter table '.$table.' '.$column)->all();
+    }
+
+    /**
+     * Compile a drop primary key command.
+     *
+     * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
+     * @param  \Illuminate\Support\Fluent  $command
+     * @return string
+     */
+    public function compileDropPrimary(Blueprint $blueprint, Fluent $command)
+    {
+        // Handled on table alteration...
     }
 
     /**
@@ -496,6 +511,22 @@ class SQLiteGrammar extends Grammar
     public function compileDropSpatialIndex(Blueprint $blueprint, Fluent $command)
     {
         throw new RuntimeException('The database driver in use does not support spatial indexes.');
+    }
+
+    /**
+     * Compile a drop foreign key command.
+     *
+     * @param  \Illuminate\Database\Schema\Blueprint  $blueprint
+     * @param  \Illuminate\Support\Fluent  $command
+     * @return array
+     */
+    public function compileDropForeign(Blueprint $blueprint, Fluent $command)
+    {
+        if (empty($command->columns)) {
+            throw new RuntimeException('This database driver does not support dropping foreign keys by name.');
+        }
+
+        // Handled on table alteration...
     }
 
     /**
@@ -560,7 +591,7 @@ class SQLiteGrammar extends Grammar
      */
     public function compileEnableForeignKeyConstraints()
     {
-        return 'PRAGMA foreign_keys = ON;';
+        return $this->pragma('foreign_keys', 'ON');
     }
 
     /**
@@ -570,7 +601,40 @@ class SQLiteGrammar extends Grammar
      */
     public function compileDisableForeignKeyConstraints()
     {
-        return 'PRAGMA foreign_keys = OFF;';
+        return $this->pragma('foreign_keys', 'OFF');
+    }
+
+    /**
+     * Compile the command to set the busy timeout.
+     *
+     * @param  int  $milliseconds
+     * @return string
+     */
+    public function compileSetBusyTimeout($milliseconds)
+    {
+        return $this->pragma('busy_timeout', $milliseconds);
+    }
+
+    /**
+     * Compile the command to set the journal mode.
+     *
+     * @param  string  $mode
+     * @return string
+     */
+    public function compileSetJournalMode($mode)
+    {
+        return $this->pragma('journal_mode', $mode);
+    }
+
+    /**
+     * Compile the command to set the synchronous mode.
+     *
+     * @param  string  $mode
+     * @return string
+     */
+    public function compileSetSynchronous($mode)
+    {
+        return $this->pragma('synchronous', $mode);
     }
 
     /**
@@ -580,7 +644,7 @@ class SQLiteGrammar extends Grammar
      */
     public function compileEnableWriteableSchema()
     {
-        return 'PRAGMA writable_schema = 1;';
+        return $this->pragma('writable_schema', 1);
     }
 
     /**
@@ -590,7 +654,19 @@ class SQLiteGrammar extends Grammar
      */
     public function compileDisableWriteableSchema()
     {
-        return 'PRAGMA writable_schema = 0;';
+        return $this->pragma('writable_schema', 0);
+    }
+
+    /**
+     * Get the SQL to set a PRAGMA value.
+     *
+     * @param  string  $name
+     * @param  mixed  $value
+     * @return string
+     */
+    protected function pragma(string $name, mixed $value): string
+    {
+        return sprintf('PRAGMA %s = %s;', $name, $value);
     }
 
     /**
